refactor: move Producer.Filter logic to its own file
